+++ Todoary-iOS/Todoary/API/Auth/AutoLoginDataManager/AutoLoginModel.swift
