+++ Todoary-iOS/Todoary/API/Auth/EmailCheckDataManager/EmailCheckDataManager.swift
@@ -10,13 +10,14 @@ import Alamofire
 class EmailCheckDataManager{
     
     func emailCheckDataManager(_ viewController: PwFindViewController, email: String){
-        AF.request("http://todoary.com:9000/auth/email/existence", method: .get, parameters: ["email":email], encoding: URLEncoding.queryString).validate().responseDecodable(of: EmailCheckModel.self) { response in
+        AF.request("https://todoary.com/auth/email/existence", method: .get, parameters: ["email":email], encoding: URLEncoding.queryString).validate().responseDecodable(of: EmailCheckModel.self) { response in
             switch response.result {
             case .success(let result):
-                if result.isSuccess{
+                switch result.code{
+                case 1000:
+                    viewController.checkEmail(1000)
                     print("비밀번호찾기 이메일 인증 성공")
-                    viewController.checkEmail(result.code)
-                }else{
+                default:
                     viewController.checkEmail(result.code)
                     print(result.message)
                 }
