+++ Todoary-iOS/Todoary/API/Auth/EmailCheckDataManager/EmailCheckModel.swift
