+++ Todoary-iOS/Todoary/API/Auth/LoginDataManager/LoginInput.swift
