+++ Todoary-iOS/Todoary/API/Auth/SignoutDataManager/SignoutDataManager.swift
@@ -13,7 +13,7 @@ class SignoutDataManager {
         let headers : HTTPHeaders = [.authorization(UserDefaults.standard.string(forKey: "accessToken")!)]
         
         //통신
-    func signout(_ viewController : UIViewController) {AF.request("http://todoary.com:9000/users/signout", method: .post, parameters: nil, headers: headers).validate().responseDecodable(of: SignoutModel.self) {response in
+    func signout(_ viewController : UIViewController) {AF.request("https://todoary.com/users/signout", method: .post, parameters: nil, headers: headers).validate().responseDecodable(of: SignoutModel.self) {response in
             switch response.result {
             case .success(let result) :
                 switch result.code {
@@ -30,9 +30,9 @@ class SignoutDataManager {
                     }
                     print("로그아웃 성공")
                 
-//                case 4000 :
-//                    let alert = DataBaseErrorAlert()
-//                    vc.present(alert, animated: true, completion: nil)
+                case 4000 :
+                    let alert = DataBaseErrorAlert()
+                    viewController.present(alert, animated: true, completion: nil)
                 default:
                     print(result.message)
                 }
