+++ Todoary-iOS/Todoary/API/Auth/SignoutDataManager/SignoutModel.swift
