+++ Todoary-iOS/Todoary/API/User/GetProfileDataManager/GetProfileInput.swift
