+++ Todoary-iOS/Todoary/API/User/ProfileDataManager/ProfileDataManager.swift
@@ -12,18 +12,18 @@ class ProfileDataManager {
     let headers : HTTPHeaders = [.authorization(UserDefaults.standard.string(forKey: "accessToken")!)]
     
     func profileDataManager( _ viewController : ProfileViewController , _ parameter: ProfileInput) {
-        AF.request("http://todoary.com:9000/users/profile", method: .patch, parameters: parameter,  encoder: JSONParameterEncoder.default , headers: headers).validate().responseDecodable(of: ProfileModel.self) { response in
+        AF.request("https://todoary.com/users/profile", method: .patch, parameters: parameter,  encoder: JSONParameterEncoder.default , headers: headers).validate().responseDecodable(of: ProfileModel.self) { response in
             switch response.result {
             case .success(let result):
-                if result.isSuccess {
-                    if result.code == 1000{
-                        print("프로필수정성공")
-                    }else{
-                        print(result.message)
-                    }
-                }else {
+                switch result.code {
+                case 1000:
+                    print("프로필수정성공")
+                case 2201:
+                    print("중복된 닉네임입니다")
+                default:
                     print(result.message)
                 }
+
             case .failure(let error):
                 print(error.localizedDescription)
             }
